# lib/api-zod/src/generated/types/adminStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStatus {
  online: boolean;
  proxyStatusCode?: number | null;
  proxyUrl?: string;
  error?: string | null;
  modelCount: number;
  accountCount: number;
  keyCount: number;
  dataDir: string;
}