# lib/api-zod/src/generated/types/jetbrainsAccount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JetbrainsAccount {
  jwt?: string | null;
  licenseId?: string | null;
  authorization?: string | null;
  has_quota?: boolean | null;
  last_updated?: number | null;
  last_quota_check?: number | null;
}