# lib/api-zod/src/generated/types/logsResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LogsResult {
  lines: string[];
  total?: number;
  file?: string;
  note?: string;
}