# lib/api-zod/src/generated/types/modelsConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ModelsConfigAnthropicModelMappings } from "./modelsConfigAnthropicModelMappings";

export interface ModelsConfig {
  models: string[];
  anthropic_model_mappings: ModelsConfigAnthropicModelMappings;
}