# lib/api-zod/src/generated/types/proxyTestResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProxyTestResult {
  status: number;
  data?: unknown | null;
  ok: boolean;
  error?: string | null;
}