# lib/api-zod/src/generated/types/testChatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TestChatRequestMessagesItem } from "./testChatRequestMessagesItem";

export interface TestChatRequest {
  apiKey: string;
  model: string;
  messages: TestChatRequestMessagesItem[];
  stream?: boolean;
  baseUrl?: string | null;
}